+++ webpack.config.js
@@ -4,6 +4,7 @@ const UglifyJsPlugin = require('uglifyjs-webpack-plugin')
 
 var package = require('./package.json');
 var loader = path.join(__dirname, "./loader");
+
 var minify = new UglifyJsPlugin({
 	uglifyOptions: {
 		ecma: 6,
@@ -19,7 +20,7 @@ var minify = new UglifyJsPlugin({
 
 module.exports = [{
 	module: {
-		rules: [{test: /.imba$/, loader: loader, options: {es5: true}}]
+		rules: [{test: /\.imba$/, loader: loader, options: {es5: true}}]
 	},
 	resolve: {extensions: ['*', '.imba', '.js']},
 	entry: "./src/imba/index.imba",
@@ -36,7 +37,7 @@ module.exports = [{
 	plugins: [minify]
 },{
 	module: {
-		rules: [{test: /.imba$/, loader: loader, options: {es5: true}}]
+		rules: [{test: /\.imba$/, loader: loader, options: {es5: true}}]
 	},
 	resolve: {extensions: ['*', '.imba', '.js']},
 	entry: "./src/compiler/compiler.imba",
@@ -49,7 +50,7 @@ module.exports = [{
 	plugins: [minify]
 },{
 	module: {
-		rules: [{test: /.imba$/, loader: loader}]
+		rules: [{test: /\.imba$/, loader: loader}]
 	},
 	resolve: {extensions: ['*', '.imba', '.js']},
 	entry: "./test/index.imba",